Replace first and follow functions with functions that utilize item_map
