+++ source/typescript/utilities/closure.ts
@@ -1,8 +1,5 @@
 import { getAccompanyingItems, Item, itemsToProductions } from "./item.js";
-
-import { EOF_SYM, Grammar } from "../types/grammar.js";
-import { TokenSymbol } from "../types/symbol";
-
+import { Grammar } from "../types/grammar.js";
 
 export function getClosure(items: Item[], grammar: Grammar): Item[] {
     let closure = [];
@@ -15,16 +12,6 @@ export function getClosure(items: Item[], grammar: Grammar): Item[] {
         closure = closure.setFilter(i => i);
 
     return closure.map(i => grammar.item_map.get(i).item);
-}
-
-export function getFollow(production_id: number, grammar: Grammar): TokenSymbol[] {
-    const prod = grammar[production_id];
-
-    return prod.bodies
-        .map(b => new Item(b.id, b.length, b.length, EOF_SYM))
-        .flatMap(i => [...grammar.item_map.get(i.id).follow.values()])
-        .setFilter()
-        .map(sym => <TokenSymbol>grammar.meta.all_symbols.get(sym));
 }
 
 /**
